% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abreviar_comunas.R
\name{abreviar_comunas}
\alias{abreviar_comunas}
\title{Abreviar comunas}
\usage{
abreviar_comunas(nombre_comuna)
}
\arguments{
\item{nombre_comuna}{Vector de nombres de comunas, como \code{\link[=comunas]{comunas()}}}
}
\value{
Vector con siglas de comunas de tres o cuatro caracteres
}
\description{
A partir de un vector de nombres de comunas (como \code{\link[=comunas]{comunas()}}), retorna nombres de comunas abreviados, como siglas de 3 caracteres (a excepción de Santiago, que queda STGO). Sirve para poner nombres de comunas donde hay poco espacio, como en mapas. Las abreviaciones no necesariamente son únicas.
}
\details{
Las siglas se construyen con las tres primeras letras de las comunas con nombres de una palabra, y en las comuna de dos o más palabras con la primera letra de la primera palabra y las dos primeras letras de la segunda palabra. Se ajustan algunos casos, como eliminar símbolos (como en O'Higgins), y sacar la "u" de las comunas que empiezan con "Que" para reducir coincidencias de abreviaciones.
}
\examples{
abreviar_comunas(c("Cerrillos", "La Florida", "Quellón"))
}
