% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actualizar_codigo_comuna.R
\name{actualizar_codigo_comuna}
\alias{actualizar_codigo_comuna}
\title{Actualizar códigos únicos territoriales comunales antiguos o desactualizados}
\usage{
actualizar_codigo_comuna(codigo_comuna)
}
\arguments{
\item{codigo_comuna}{Códigos comunales en formato numérico}
}
\value{
Códigos comunales actualizados correspondientes a los desactualizados entregados. Si los códigos entregados no son antiguos, se mantienen sin cambios.
}
\description{
Función que recibe códigos únicos territoriales desactualizados (confirmar con \code{\link[=is_codigo_comuna_antiguo]{is_codigo_comuna_antiguo()}}), y retorna las versiones actualizadas al 2017 (posterior a la creación de la región de Ñuble).
}
\examples{
actualizar_codigo_comuna(8408)

}
