% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregar_clasificacion.R
\name{agregar_clasificacion}
\alias{agregar_clasificacion}
\title{Agregar clasificación comunal PNDR a comunas}
\usage{
agregar_clasificacion(codigo_comuna)
}
\arguments{
\item{codigo_comuna}{Vector de códigos de comuna, idealmente \code{codigo_comuna}}
}
\value{
Vector con clasificación comunal (Rural, Mixta o Urbana)
}
\description{
Para un vector de comunas, entrega la clasificación comunal PNDR correspondiente (Rural, Mixta o Urbana), tal como aparecen en \link{clasificacion}.
}
\examples{
agregar_clasificacion(8107)
}
