% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregar_orden_region.R
\name{agregar_orden_region}
\alias{agregar_orden_region}
\title{Agregar orden geográfico (norte a sur) a códigos de regiones}
\usage{
agregar_orden_region(codigo_region)
}
\arguments{
\item{codigo_region}{Vector de códigos únicos de región (entre 1 y 16)}
}
\value{
Vector con la posición de norte a sur de la región respectiva
}
\description{
Función usada internamente para \code{\link[=ordenar_regiones]{ordenar_regiones()}}. Recibe un vector de números de región (entre 1 y 16) y retorna la posición de norte a sur de dicha región, de modo que Arica sea 1 y Magallanes sea 16.
}
\examples{
agregar_orden_region(15) # Arica
}
