% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articulo_region.R
\name{articulo_region}
\alias{articulo_region}
\title{Artículos (de/del) de cada región de Chile}
\usage{
articulo_region(nombre_region)
}
\arguments{
\item{nombre_region}{Nombres de región, como aparecen en \code{\link[=regiones]{regiones()}}}
}
\value{
Vector de artículos para cada nombre de región. Retorna NA si no se detecta la región.
}
\description{
Esta función entrega el artículo usado para cada región de Chile; por ejemplo, Valparaíso es la "Región \emph{de} Valparaíso", pero Maule es la "Región \emph{del} Maule".
}
\examples{
articulo_region("Ñuble")
}
