% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.codigo_comuna.R
\name{as.codigo_comuna}
\alias{as.codigo_comuna}
\title{Convertir nombres de comunas a códigos de comuna}
\usage{
as.codigo_comuna(nombres_comunas)
}
\arguments{
\item{nombres_comunas}{Nombres de comuna (como los que aparecen en \code{comunal::comunas()})}
}
\value{
Vector numérico con códigos de comuna
}
\description{
Entregando nombres de comuna correctos (como los que aparecen en \code{comunal::comunas()} o en \code{comunal::territorios}), retorna los códigos de comuna correspondientes en formato numérico. Retorna NA si no corresponde con ninguna.
}
