% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos_comunas.R
\name{codigos_comunas}
\alias{codigos_comunas}
\title{Códigos de comunas de Chile}
\usage{
codigos_comunas()
}
\value{
Vector de códigos de comunas de Chile
}
\description{
Vector numérico que contiene los códigos únicos territoriales de las 346 comunas de Chile, tal como aparecen en la tabla \link{territorios}.
}
\examples{
codigos_comunas()

}
