% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comunas.R
\name{comunas}
\alias{comunas}
\title{Nombres de comunas de Chile}
\usage{
comunas()
}
\value{
Vector de nombres de comunas de Chile
}
\description{
Vector de texto que contiene el nombre de las 346 comunas de Chile, tal como aparecen en la tabla \link{territorios}.
}
\examples{
comunas()

}
