% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contar_comunas.R
\name{contar_comunas}
\alias{contar_comunas}
\title{Conteo de comunas disponibles, indicando las faltantes}
\usage{
contar_comunas(datos, variable = NULL, largo = 30)
}
\arguments{
\item{datos}{Dataframe con una columna de nombre de comunas, o vector de nombres de comunas}

\item{variable}{Columna del dataframe con los nombres de comunas (se pasa sin comillas, p.ej. \code{comuna})}

\item{largo}{Si faltan comunas en los datos, enumera esta cantidad de comunas. Por defecto son 30, y el argumento se pasa a \code{\link[=redactar_comunas]{redactar_comunas()}}.}
}
\value{
Dataframe o vector intacto pero en modo invisible, con mensajes de conteo de comunas
}
\description{
Revisa un vector o una tabla de datos (asumiendo la columna \code{nombre_comuna}, o indicando una columna que contenga nombres de comunas) e indica el conteo de comunas únicas, y emite mensajes dependiendo de esta cantidad: si es correcta (345 o 346), si es menor (indicando las que faltan) o si es mayor (recomendando validar con \code{\link[=validar_comunas]{validar_comunas()}}).
}
\details{
Sirve para revisar rápidamente que los datos abarquen todos los territorios del país, o indicar si faltan algunos.
}
\examples{
territorios |>
  dplyr::slice_sample(n = 300) |>
  contar_comunas()
}
