% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estandariza_regiones.R
\name{estandariza_regiones}
\alias{estandariza_regiones}
\title{Estandariza nombres de regiones de Chile}
\usage{
estandariza_regiones(regiones)
}
\arguments{
\item{regiones}{Un vector de caracteres}
}
\value{
Devuelve un vector de caracteres con los nombres
oficiales de las regiones de Chile
}
\description{
Esta función reconoce nombres de regiones de
Chile en un vector de caracteres y devuelve el nombre oficial
de las mismas (por ejemplo, "RM" lo convierte en "Metropolitana
de Santiago"). Si no reconoce un nombre de región, entonces
mantiene el provisto originalmente en en el vector de caracteres.
La función está pensada para ser utilizada en conjunto
con funciones como dplyr::mutate.
}
\examples{
regiones <- c("Aisén", "Aysén", "Región del Bio-bío",
"Biobio", "RM", "Metropolitana")
estandariza_regiones(regiones)

}
