% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.comuna.R
\name{is.comuna}
\alias{is.comuna}
\title{Evaluar si es una comuna}
\usage{
is.comuna(x)
}
\arguments{
\item{x}{Elemento/s a evaluar}
}
\value{
TRUE o FALSE si es o no es una comuna
}
\description{
Evaluar si es una comuna
}
\examples{
is.comuna("Panguipulli")
}
