% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_codigo_comuna_antiguo.R
\name{is_codigo_comuna_antiguo}
\alias{is_codigo_comuna_antiguo}
\title{Evaluar si un código único territorial comunal es antiguo o desactualizado}
\usage{
is_codigo_comuna_antiguo(codigo_comuna)
}
\arguments{
\item{codigo_comuna}{Códigos comunales en formato numérico}
}
\value{
Retorna TRUE o FALSE si es o no es un código único territorial antiguo (ver \link{cut_historicos})
}
\description{
Evaluar si un código único territorial comunal es antiguo o desactualizado
}
\examples{
is_codigo_comuna_antiguo(8408)

}
