% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_nombre_region.R
\name{is_nombre_region}
\alias{is_nombre_region}
\title{Evaluar si un texto corresponde a un nombre de región de Chile válido}
\usage{
is_nombre_region(nombre_region)
}
\arguments{
\item{codigo_comuna}{Elemento/s a evaluar}
}
\value{
TRUE o FALSE si es o no es un código único territorial válido
}
\description{
Dado un vector de cualquier largo, retorna TRUE o FALSE para cada elemento de acuerdo si se corresponde con los nombres de regiones de Chile, disponibles en \link{territorios}.
}
\examples{
is_nombre_region("Maule")
}
