% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redactar_region.R
\name{redactar_region}
\alias{redactar_region}
\title{Redactar nombres de regiones de Chile}
\usage{
redactar_region(nombre_region)
}
\arguments{
\item{nombre_region}{Nombres de región, como aparecen en \code{\link[=regiones]{regiones()}}}
}
\value{
Vector de nombres de región redactados, incluyendo la palabra "Región", y el artículo apropiado (de o del, según \code{\link[=articulo_region]{articulo_region()}}) para cada nombre de región.
}
\description{
A partir de los nombres de regiones de Chile (como aparecen en \code{\link[=regiones]{regiones()}}), retorna la redacción de los nombres incluyendo "Región de/del"; por ejemplo, "Maule" se vuelve en "Región del Maule".
}
\examples{
redactar_region("Maule")
}
