% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regiones.R
\name{regiones}
\alias{regiones}
\title{Nombres de regiones de Chile}
\usage{
regiones()
}
\value{
Vector de nombres de regiones de Chile
}
\description{
Vector de texto que contiene el nombre de las 15 regiones de Chile, ordenadas de norte a sur, tal como aparecen en la tabla \link{territorios}.
}
\examples{
regiones()

}
