% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/territorios.R
\docType{data}
\name{territorios}
\alias{territorios}
\title{Códigos Únicos Territoriales}
\format{
A data frame with 346 rows and 6 columns:
\describe{
\item{codigo_region}{Código único territorial de las regiones de Chile (número entre 1 y 16)}
\item{nombre_region}{Nombre de las regiones de Chile}
\item{codigo_provincia}{Código único territorial de las provincias de Chile}
\item{nombre_provincia}{Nombre de las provincias de Chile}
\item{codigo_comuna}{Código único territorial de las comunas de Chile}
\item{nombre_comuna}{Nombre de las comunas de Chile}
}
}
\source{
\url{https://www.subdere.gov.cl/documentacion/códigos-únicos-territoriales-actualizados-al-06-de-septiembre-2018}
}
\usage{
territorios
}
\description{
Por Decreto Exento N° 1.115, del Ministerio del Interior y Seguridad Pública, publicado en el Diario Oficial el 21 de septiembre de 2018, se establecen las abreviaturas para identificar las regiones del país, unificando además los códigos territoriales (incluyendo la región de Ñuble, sus provincias y comunas).
}
\keyword{datasets}
