% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubicar_comunas.R
\name{ubicar_comunas}
\alias{ubicar_comunas}
\title{Ubicar comunas en la región que les corresponde}
\usage{
ubicar_comunas(nombre_comuna = NULL, codigo_comuna = NULL)
}
\arguments{
\item{nombre_comuna}{Nombres de comunas, como aparecen en \code{\link[=comunas]{comunas()}}}

\item{codigo_comuna}{Códigos comunales, como aparecen en \link{territorios}}
}
\value{
Vector con nombres de regiones correspondientes
}
\description{
Para un vector de nombres de comuna o códigos comunales, retorna las regiones donde se ubican dichas regiones.
}
\examples{
ubicar_comunas(c("Cerrillos", "Navidad"))
}
