% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validar_regiones.R
\name{validar_regiones}
\alias{validar_regiones}
\title{Validación de calidad de nombres de regiones de Chile}
\usage{
validar_regiones(nombre_region)
}
\arguments{
\item{nombre_region}{Columna con nombres de regiones}
}
\value{
Dataframe intacto, mensajes de diagnóstico
}
\description{
Esta función recibe la columna con nombres de regiones de un dataframe (idealmente \code{nombre_region}), y retorna una evaluación de posibles problemas con los nombres existentes.
}
