% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nombre_comuna.R
\name{as.nombre_comuna}
\alias{as.nombre_comuna}
\title{Convertir códigos comunales en nombres de comunas}
\usage{
as.nombre_comuna(codigos_comunas)
}
\arguments{
\item{codigos_comunas}{Códigos comunales en formato numérico}
}
\value{
Vector con nombres de comuna
}
\description{
Entregando códigos comunales (como los que aparecen en \code{comunal::territorios}), retorna los nombres de comuna correspondientes. Retorna NA si no corresponde con ninguna.
}
