% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.comuna.R
\name{is.comuna}
\alias{is.comuna}
\title{Evaluar si un texto corresponde al nombre válido de una comuna de Chile}
\usage{
is.comuna(x)
}
\arguments{
\item{x}{Elemento/s a evaluar}
}
\value{
TRUE o FALSE si es o no es una comuna válida
}
\description{
Dado un vector de cualquier largo, retorna TRUE o FALSE para cada elemento de acuerdo si se corresponde con los nombres de comunas oficiales, disponibles en la función \code{\link[=comunas]{comunas()}}.
}
\examples{
is.comuna("Panguipulli")

}
