% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_comunas.R
\name{limpiar_comunas}
\alias{limpiar_comunas}
\title{Limpieza de nombres de comunas de Chile a sus nombres oficiales}
\usage{
limpiar_comunas(nombre_comuna, aproximar = TRUE, mostrar_proceso = TRUE)
}
\arguments{
\item{nombre_comuna}{Vector de nombres de comunas}

\item{aproximar}{El paso de limpieza por aproximación y coincidencia de nombres puede entregar resultados inexactos. Cambiar a FALSE para omitir.}

\item{mostrar_proceso}{Por defecto, muestra una tabla con el resultado del proceso de limpieza. Cambiar a FALSE para ocultar.}
}
\value{
Vector de nombres de comunas con correcciones aplicadas.
}
\description{
A partir de un vector de nombres de columnas, se realizan tres pasos de limpieza (confirmación de nombres correctas, limpieza de texto, detección por coincidencia) para retornar los nombres de comunas oficiales apropiados. Los nombres de comunas son los que aparecen en \code{\link[=comunas]{comunas()}}.
}
\examples{
limpiar_comunas(c("COLCHANE", "Alto Ospicio", "probidencia", "huara", "laflorida", "cerritos", "llay-llay"))
}
