% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redactar_comunas.R
\name{redactar_comunas}
\alias{redactar_comunas}
\title{Redactar una secuencia de comunas en una oración}
\usage{
redactar_comunas(x)
}
\arguments{
\item{x}{Nombres de comunas}
}
\value{
Texto con comunas separadas por comas
}
\description{
Para un vector de nombrees de comuna, retorna un texto con las comunas redactadas y separadas por comas.
}
\examples{
redactar_comunas(c("Paine", "Isla de Maipo", "Pirque"))
}
