% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/territorios.R
\docType{data}
\name{territorios}
\alias{territorios}
\title{Códigos Únicos Territoriales}
\format{
A data frame with 346 rows and 6 columns:
\describe{
\item{codigo_region}{Describir}
\item{nombre_region, iso3}{Describir}
\item{codigo_provincia}{Describir}
\item{nombre_provincia}{Describir}
\item{codigo_comuna}{Describir}
\item{nombre_comuna}{Describir}
}
}
\source{
\url{https://www.subdere.gov.cl/documentacion/códigos-únicos-territoriales-actualizados-al-06-de-septiembre-2018}
}
\usage{
territorios
}
\description{
Por Decreto Exento N° 1.115, del Ministerio del Interior y Seguridad Pública, publicado en el Diario Oficial el 21 de septiembre de 2018, se establecen las abreviaturas para identificar las regiones del país, unificando además los códigos territoriales (incluyendo la región de Ñuble, sus provincias y comunas).
}
\keyword{datasets}
