% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validar_comunas.R
\name{validar_comunas}
\alias{validar_comunas}
\title{Validación de calidad de nombres de comunas de Chile}
\usage{
validar_comunas(datos, variable = "nombre_comuna")
}
\arguments{
\item{nombre_comuna}{Columna con nombres de comunas, o vector con nombres de comunas}
}
\value{
Dataframe o vector intacto, con mensajes de diagnóstico si se encuentran problemas de calidad
}
\description{
Esta función recibe la columna con nombres de comunas de un dataframe (idealmente \code{nombre_comuna}), y retorna una evaluación de posibles problemas con los nombres existentes. Funciona tanto con un dataframe con una columna \code{nombre_comuna}, o un vector que contenga los nombres de comunas a evaluar.
}
\examples{
validar_comunas(c("colliguay", "la florida", "paine"))

}
